+++ src/main/java/shop/tryit/domain/payment/dto/PaymentResponseDto.java
@@ -1,4 +1,4 @@
-package shop.tryit.web.payment;
+package shop.tryit.domain.payment.dto;
 
 import lombok.Builder;
 import lombok.Data;
